[Backport] do not test anymore LF 1.6 and 1.7 (#17434)
