--- language-support/java/codegen/src/test/java/com/digitalasset/testing/AllTestsFor1_6.java
@@ -1,14 +1,0 @@
-// Copyright (c) 2023 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-package com.daml.testing;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Suite;
-
-@RunWith(Suite.class)
-@Suite.SuiteClasses({
-  DecimalTestForAll.class,
-  EnumTestForForAll.class,
-})
-public class AllTestsFor1_6 {}
