+++ language-support/java/codegen/src/test/java/com/digitalasset/testing/AllTestsFor1_dev.java
@@ -10,7 +10,7 @@ import org.junit.runners.Suite;
 @Suite.SuiteClasses({
   DecimalTestForAll.class,
   EnumTestForForAll.class,
-  NumericTestFor1_7AndFor1_8AndFor1_11AndFor1_12ndFor1_13AndFor1_14AndFor1_15AndFor1_dev.class,
+  NumericTestForAll.class,
   GenMapTestFor1_11AndFor1_12ndFor1_13AndFor1_14AndFor1_15AndFor1_dev.class,
 })
 public class AllTestsFor1_dev {}
